+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+  * Add new non-openssl DTLS under `dimpl` feature #708
   * Fix bug when TWCC sequence starts above 32768 #717
 
 # 0.11.1